+++ src/event_queue.rs
@@ -1,8 +1,9 @@
+#[cfg(not(loom))]
 #[cfg(test)]
 mod test;
 
 use crate::sync::{Ordering, AtomicUsize};
-use crate::sync::{Mutex, MutexGuard, Arc};
+use crate::sync::{Mutex, Arc};
 use crate::sync::{SpinMutex};
 
 use std::ptr::{null_mut, null};
@@ -15,7 +16,7 @@ use crate::cursor::Cursor;
 use crate::dynamic_chunk::{DynamicChunk};
 #[cfg(feature = "double_buffering")]
 use crate::dynamic_chunk::{DynamicChunkRecycled};
-use crate::StartPositionEpoch;
+use crate::{StartPositionEpoch};
 
 /// This way you can control when chunk's memory deallocation happens.
 #[derive(PartialEq)]
@@ -23,6 +24,8 @@ pub enum CleanupMode{
     /// Cleanup will be called when chunk fully read.
     ///
     /// In this mode memory will be freed ASAP - right in the end of reader consumption session.
+    ///
+    /// !! Not allowed for spmc !!
     OnChunkRead,
     /// Cleanup will be called when new chunk created.
     OnNewChunk,
@@ -33,17 +36,15 @@ pub enum CleanupMode{
 pub trait Settings{
     const MIN_CHUNK_SIZE : u32;
     const MAX_CHUNK_SIZE : u32;
-    const CLEANUP: CleanupMode;
+    const CLEANUP        : CleanupMode;
+
+    /// Lock on new chunk cleanup event. Will dead-lock if already locked.
+    const LOCK_ON_NEW_CHUNK_CLEANUP: bool;
+    /// Call cleanup on unsubscribe?
+    const CLEANUP_IN_UNSUBSCRIBE: bool;
 }
 
-pub struct DefaultSettings{}
-impl Settings for DefaultSettings{
-    const MIN_CHUNK_SIZE: u32 = 4;
-    const MAX_CHUNK_SIZE: u32 = u32::MAX / 4;
-    const CLEANUP: CleanupMode = CleanupMode::OnChunkRead;
-}
-
-struct List<T, S: Settings>{
+pub struct List<T, S: Settings>{
     first: *mut DynamicChunk<T, S>,
     last : *mut DynamicChunk<T, S>,
     chunk_id_counter: usize,
@@ -56,8 +57,8 @@ struct List<T, S: Settings>{
     free_chunk: Option<DynamicChunkRecycled<T, S>>,
 }
 
-pub struct EventQueue<T, S: Settings = DefaultSettings>{
-    list  : Mutex<List<T, S>>,
+pub struct EventQueue<T, S: Settings>{
+    pub(crate) list  : Mutex<List<T, S>>,
 
     /// All atomic op relaxed. Just to speed up `try_clean` check (opportunistic check).
     /// Mutated under list lock.
@@ -70,13 +71,14 @@ pub struct EventQueue<T, S: Settings = DefaultSettings>{
     _pinned: PhantomPinned,
 }
 
-unsafe impl<T, S: Settings> Send for EventQueue<T, S>{}
-unsafe impl<T, S: Settings> Sync for EventQueue<T, S>{}
-
+//unsafe impl<T, S: Settings> Send for EventQueue<T, S>{}
+//unsafe impl<T, S: Settings> Sync for EventQueue<T, S>{}
 
 impl<T, S: Settings> EventQueue<T, S>
 {
-    pub fn new() -> Pin<Arc<Self>>{
+    pub fn with_capacity(new_capacity: u32) -> Pin<Arc<Self>>{
+        assert!(S::MIN_CHUNK_SIZE <= new_capacity && new_capacity <= S::MAX_CHUNK_SIZE);
+
         let this = Arc::pin(Self{
             list: Mutex::new(List{
                 first: null_mut(),
@@ -94,7 +96,7 @@ impl<T, S: Settings> EventQueue<T, S>
         });
 
         let node = DynamicChunk::<T, S>::construct(
-            0, StartPositionEpoch::zero(), &*this, S::MIN_CHUNK_SIZE as usize);
+            0, StartPositionEpoch::zero(), &*this, new_capacity as usize);
 
         unsafe {
             let event = &mut *(&*this as *const _ as *mut EventQueue<T, S>);
@@ -122,8 +124,8 @@ impl<T, S: Settings> EventQueue<T, S>
             let mut new_node: *mut DynamicChunk<T, S> = null_mut();
 
             if let Some(recycled_chunk) = &list.free_chunk {
-                // Check if recycled_chunk have sufficient capacity. We never go down in capacity.
-                if recycled_chunk.capacity() >= size {
+                // Check if recycled_chunk have exact capacity.
+                if recycled_chunk.capacity() == size {
                     // unwrap_unchecked()
                     new_node =
                     match list.free_chunk.take() {
@@ -134,6 +136,9 @@ impl<T, S: Settings> EventQueue<T, S>
                                 epoch) }
                         }, None => unsafe { std::hint::unreachable_unchecked() },
                     }
+                } else {
+                    // TODO: try free in cleanup somehow
+                    list.free_chunk = None;
                 }
             }
 
@@ -152,12 +157,23 @@ impl<T, S: Settings> EventQueue<T, S>
     }
 
     #[inline]
+    fn on_new_chunk_cleanup(&self, list: &mut List<T, S>){
+        if S::CLEANUP == CleanupMode::OnNewChunk{
+            // this should acts as compile-time-if.
+            if S::LOCK_ON_NEW_CHUNK_CLEANUP{
+                // `cleanup` - locks internally
+                self.cleanup();
+            } else {
+                self.cleanup_impl(list);
+            }
+        }
+    }
+
+    #[inline]
     fn add_chunk(&self, list: &mut List<T, S>) -> &mut DynamicChunk<T, S>{
         let node = unsafe{&mut *list.last};
 
-        if S::CLEANUP == CleanupMode::OnNewChunk{
-            self.cleanup_impl(list);
-        }
+        self.on_new_chunk_cleanup(list);
 
         // Size pattern 4,4,8,8,16,16
         let new_size: usize = {
@@ -171,28 +187,26 @@ impl<T, S: Settings> EventQueue<T, S>
         self.add_chunk_sized(list, new_size)
     }
 
-    // Leave this for a while. Have feeling that this one should be faster.
-    // #[inline]
-    // pub fn push(&self, value: T){
-    //     let mut list = self.list.lock();
-    //     let mut node = unsafe{&mut *list.last};
-    //
-    //     // Relaxed because we update only under lock
-    //     let len_and_epoch: LenAndEpoch = node.storage.len_and_epoch(Ordering::Relaxed);
-    //     let mut storage_len = len_and_epoch.len();
-    //     let epoch = len_and_epoch.epoch();
-    //
-    //     if /*unlikely*/ storage_len as usize == CHUNK_SIZE{
-    //         node = self.add_chunk(&mut *list);
-    //         storage_len = 0;
-    //     }
-    //
-    //     unsafe { node.storage.push_at(value, storage_len, epoch, Ordering::Release); }
-    // }
-
+    // Have 10% better performance. Observable in spmc.
     #[inline]
-    pub fn push(&self, value: T){
-        let mut list = self.list.lock();
+    pub fn push(&self, list: &mut List<T, S>, value: T){
+        let mut node = unsafe{&mut *list.last};
+
+        // Relaxed because we update only under lock
+        let chunk_state = node.chunk_state(Ordering::Relaxed);
+        let mut storage_len = chunk_state.len();
+
+        if /*unlikely*/ storage_len == node.capacity() as u32{
+            node = self.add_chunk(&mut *list);
+            storage_len = 0;
+        }
+
+        unsafe { node.push_at(value, storage_len, chunk_state, Ordering::Release); }
+    }
+
+/*
+    #[inline]
+    pub fn push(&self, list: &mut List<T, S>, value: T){
         let node = unsafe{&mut *list.last};
 
         if let Err(err) = node.try_push(value, Ordering::Release){
@@ -202,13 +216,13 @@ impl<T, S: Settings> EventQueue<T, S>
             }
         }
     }
+*/
 
     // Not an Extend trait, because Extend::extend(&mut self)
     #[inline]
-    pub fn extend<I>(&self, iter: I)
+    pub fn extend<I>(&self, list: &mut List<T, S>, iter: I)
         where I: IntoIterator<Item = T>
     {
-        let mut list = self.list.lock();
         let mut node = unsafe{&mut *list.last};
 
         let mut iter = iter.into_iter();
@@ -227,9 +241,7 @@ impl<T, S: Settings> EventQueue<T, S>
 
     /// EventReader will start receive events from NOW.
     /// It will not see events that was pushed BEFORE subscription.
-    pub fn subscribe(&self) -> EventReader<T, S>{
-        let list = self.list.lock();
-
+    pub fn subscribe(&self, list: &mut List<T, S>) -> EventReader<T, S>{
         let prev_readers = self.readers.fetch_add(1, Ordering::Relaxed);
         if prev_readers == 0{
             // Keep alive. Decrements in unsubscribe
@@ -277,7 +289,7 @@ impl<T, S: Settings> EventQueue<T, S>
 
         let prev_readers = self.readers.fetch_sub(1, Ordering::Relaxed);
 
-        if S::CLEANUP != CleanupMode::Never{
+        if S::CLEANUP_IN_UNSUBSCRIBE && S::CLEANUP != CleanupMode::Never{
             self.cleanup_impl(&mut *list);
         }
 
@@ -336,9 +348,6 @@ impl<T, S: Settings> EventQueue<T, S>
         }
     }
 
-    /// Free all completely read chunks.
-    ///
-    /// Called automatically with [Settings::CLEANUP] != Never.
     pub fn cleanup(&self){
         self.cleanup_impl(&mut *self.list.lock());
     }
@@ -346,7 +355,7 @@ impl<T, S: Settings> EventQueue<T, S>
     #[inline]
     fn set_start_position(
         &self,
-        list: &mut MutexGuard<List<T, S>>,
+        list: &mut List<T, S>,
         new_start_position: Cursor<T, S>)
     {
         *self.start_position.lock() = new_start_position;
@@ -366,40 +375,17 @@ impl<T, S: Settings> EventQueue<T, S>
         }
     }
 
-
-    /// "Lazily move" all readers positions to the "end of the queue". From readers perspective,
-    /// equivalent to conventional `clear`.
-    ///
-    /// Does not free memory by itself - all readers need to be touched to free the memory.
-    ///
-    /// "End of the queue" - is the queue's end position at the moment of the `clear` call.
-    ///
-    /// "Lazy move" - means that reader actually change position and mark passed chunks,
-    /// as "read", only when actual read starts.
-    pub fn clear(&self){
-        let mut list = self.list.lock();
-
+    pub fn clear(&self, list: &mut List<T, S>){
         let last_chunk = unsafe{ &*list.last };
         let chunk_len = last_chunk.chunk_state(Ordering::Relaxed).len() as usize;
 
-        self.set_start_position(&mut list, Cursor {
+        self.set_start_position(list, Cursor {
             chunk: last_chunk,
             index: chunk_len
         });
     }
 
-
-
-    /// "Lazily move" all readers positions to the `len`-th element from the end of the queue.
-    /// From readers perspective, equivalent to conventional `truncate` from the other side.
-    ///
-    /// Does not free memory by itself - all readers need to be touched to free memory.
-    ///
-    /// "Lazy move" - means that reader actually change position and mark passed chunks
-    /// as "read", only when actual read starts.
-    pub fn truncate_front(&self, len: usize) {
-        let mut list = self.list.lock();
-
+    pub fn truncate_front(&self, list: &mut List<T, S>, len: usize) {
         // make chunks* array
         let chunks_count= unsafe {
             list.chunk_id_counter/*(*list.last).id*/ - (*list.first).id() + 1
@@ -427,7 +413,7 @@ impl<T, S: Settings> EventQueue<T, S>
             let chunk_len = chunk.chunk_state(Ordering::Relaxed).len() as usize;
             total_len += chunk_len;
             if total_len >= len{
-                self.set_start_position(&mut list, Cursor {
+                self.set_start_position(list, Cursor {
                     chunk: chunks[i],
                     index: total_len - len
                 });
@@ -435,34 +421,17 @@ impl<T, S: Settings> EventQueue<T, S>
             }
         }
 
-        // len is bigger then total_len
+        // len is bigger then total_len.
+        // do nothing.
     }
 
-    /// Adds chunk with `new_capacity` capacity. All next writes will be on new chunk.
-    ///
-    /// Use this, in conjunction with [clear](Self::clear) / [truncate_front](Self::truncate_front),
-    /// to reduce memory pressure ASAP.
-    /// Total capacity will be temporarily increased, until readers get to the new chunk.
-    pub fn change_chunk_capacity(&self, new_capacity: u32){
+    pub fn change_chunk_capacity(&self, list: &mut List<T, S>, new_capacity: u32){
         assert!(S::MIN_CHUNK_SIZE <= new_capacity && new_capacity <= S::MAX_CHUNK_SIZE);
-
-        let mut list = self.list.lock();
-        #[cfg(feature = "double_buffering")]
-        {
-            list.free_chunk = None;
-        }
+        self.on_new_chunk_cleanup(list);
         self.add_chunk_sized(&mut *list, new_capacity as usize);
+    }
 
-/*        if S::AUTO_CLEANUP {
-            if self.readers.load(Ordering::Relaxed) == 0{
-                self.cleanup_impl(&mut list);
-            }
-        }
-*/    }
-
-    /// Returns total chunks capacity.
-    pub fn total_capacity(&self) -> usize {
-        let list = self.list.lock();
+    pub fn total_capacity(&self, list: &mut List<T, S>) -> usize {
         let mut total = 0;
         unsafe {
             foreach_chunk(
@@ -477,9 +446,7 @@ impl<T, S: Settings> EventQueue<T, S>
         total
     }
 
-    /// Returns last/active chunk capacity
-    pub fn chunk_capacity(&self) -> usize {
-        let list = self.list.lock();
+    pub fn chunk_capacity(&self, list: &mut List<T, S>) -> usize {
         unsafe { (*list.last).capacity() }
     }