+++ src/sync/build.rs
@@ -14,7 +14,7 @@ impl<T> Mutex<T>{
         Self(loom::sync::Mutex::new(data))
     }
 
-    pub fn lock(&self) -> MutexGuard<'_, T> {
+    pub fn lock(&self) -> loom::sync::MutexGuard<'_, T> {
         self.0.lock().unwrap()
     }
 
@@ -28,9 +28,6 @@ impl<T> Mutex<T>{
 }
 
 #[cfg(loom)]
-pub(crate) use loom::sync::{MutexGuard};
-
-#[cfg(loom)]
 pub(crate) type SpinMutex<T> = Mutex<T>;
 
 // ==========================================================================================
@@ -42,7 +39,7 @@ pub(crate) use std::sync::atomic::{AtomicPtr, AtomicUsize, AtomicU64, Ordering};
 pub(crate) use std::sync::Arc;
 
 #[cfg(not(loom))]
-pub(crate) use parking_lot::{Mutex, MutexGuard};
+pub(crate) use parking_lot::{Mutex};
 
 #[cfg(not(loom))]
 pub(crate) use spin::mutex::{SpinMutex};