+++ src/tests/mod.rs
@@ -1,8 +1,13 @@
+#[allow(dead_code)]
 pub(crate) mod utils;
+
 mod common;
 
 #[cfg(not(loom))]
-mod test;
+mod mpmc;
+
+#[cfg(not(loom))]
+mod spmc;
 
 #[cfg(loom)]
 mod loom_test;